+++ apps/devpilot-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @svton/devpilot-web
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @svton/service@0.5.0
+
 ## 0.1.2
 
 ### Patch Changes